+++ src/Method/Payment/Create.php
@@ -4,6 +4,7 @@ namespace Cardinity\Method\Payment;
 
 use Cardinity\Method\MethodInterface;
 use Symfony\Component\Validator\Constraints as Assert;
+use Cardinity\Method\Validators\CallbackUrlConstraint;
 
 
 class Create implements MethodInterface
@@ -19,22 +20,22 @@ class Create implements MethodInterface
         $this->attributes = $attributes;
     }
 
-    public function getAction()
+    public function getAction(): string
     {
         return 'payments';
     }
 
-    public function getMethod()
+    public function getMethod(): string
     {
         return MethodInterface::POST;
     }
 
-    public function getAttributes()
+    public function getAttributes(): array
     {
         return $this->attributes;
     }
 
-    public function createResultObject()
+    public function createResultObject(): Payment
     {
         return new Payment();
     }
@@ -42,11 +43,12 @@ class Create implements MethodInterface
     public function getValidationConstraints()
     {
         return new Assert\Collection([
-            'amount' =>  $this->buildElement('float', 1),
+            'amount' => $this->buildElement('float', 1),
             'currency' => $this->buildElement('string', 1, ['min' => 3,'max' => 3]),
             'settle' => $this->buildElement('bool'),
             'order_id' => $this->buildElement('string', 0, ['min' => 2,'max' => 50]),
             'description' => $this->buildElement('string', 0, ['max' => 255]),
+            'statement_descriptor_suffix' => $this->buildElement('string', 0, ['max' => 25]),
             'country' => $this->buildElement('string', 1, ['min' => 2,'max' => 2]),
             'payment_method' => new Assert\Required([
                 new Assert\Type([
@@ -69,7 +71,7 @@ class Create implements MethodInterface
         ]);
     }
 
-    private function getPaymentInstrumentConstraints($method)
+    private function getPaymentInstrumentConstraints($method): ?object
     {
         switch ($method) {
             case self::CARD:
@@ -79,7 +81,7 @@ class Create implements MethodInterface
                         new Assert\Luhn()
                     ]),
                     'exp_year' => $this->buildElement(
-                        'integer', 1, 
+                        'integer', 1,
                         ['min' => 4,'max' => 4],
                         new Assert\Range(['min' => date('Y')])
                     ),
@@ -101,10 +103,10 @@ class Create implements MethodInterface
         );
     }
 
-    private function getThreeDS2DataConstraints()
+    private function getThreeDS2DataConstraints(): object
     {
         return new Assert\Collection([
-            'notification_url' => $this->buildElement('string', 1),
+            'notification_url' => $this->getNotificationUrlConstraints(),
             'browser_info' => $this->getBrowserInfoConstraints(),
             'billing_address' => new Assert\Optional(
                 $this->getAdressConstraints()
@@ -115,6 +117,34 @@ class Create implements MethodInterface
             'cardholder_info' => new Assert\Optional(
                 $this->getCardHolderInfoConstraints()
             ),
+        ]);
+    }
+
+    public function getNotificationUrlConstraints()
+    {
+        return new Assert\Required([
+            new Assert\NotBlank(),
+            new Assert\Type([
+                'type' => 'string',
+                'message' => 'The value {{ value }} is not a valid {{ type }}.'
+            ]),
+            new CallbackUrlConstraint(),
+            new Assert\Url([
+                'message' => 'The protocol of {{ value }} should be "http" or "https".',
+                'protocols' => ['http', 'https'],
+            ]),
+        ]);
+    }
+
+    public function getIpAddressConstraints()
+    {
+        return new Assert\Optional([
+            new Assert\NotBlank(),
+            new CallbackUrlConstraint(),
+            new Assert\Type([
+                'type' => 'string',
+                'message' => 'The value {{ value }} is not a valid {{ type }}.'
+            ]),
         ]);
     }
 
@@ -129,58 +159,68 @@ class Create implements MethodInterface
             'user_agent' => $this->buildElement('string', 1),
             'color_depth' => $this->buildElement('integer', 1),
             'time_zone' => $this->buildElement('integer', 1),
-            'ip_address' => new Assert\Optional($this->buildElement('string')),
+            'ip_address' => $this->getIpAddressConstraints(),
             'javascript_enabled' => new Assert\Optional($this->buildElement('bool')),
             'java_enabled' => new Assert\Optional($this->buildElement('bool')),
         ]);
     }
 
-    private function getAdressConstraints()
+    private function getAdressConstraints(): object
     {
         return new Assert\Collection([
-            'address_line1' => $this->buildElement('string', 1, ['max'=>50]),
+            'address_line1' => $this->buildElement('string', 1, ['max' => 50]),
             'address_line2' => new Assert\Optional(
-                $this->buildElement('string', 1, ['max'=>50])
+                $this->buildElement('string', 1, ['max' => 50])
             ),
             'address_line3' => new Assert\Optional(
-                $this->buildElement('string', 0, ['max'=>50])
+                $this->buildElement('string', 0, ['max' => 50])
             ),
-            'city' => $this->buildElement('string', 1, ['max'=>50]),
-            'country' => $this->buildElement('string', 1, ['max'=>10]),
-            'postal_code' => $this->buildElement('string', 1, ['max'=>16]),
+            'city' => $this->buildElement('string', 1, ['max' => 50]),
+            'country' => $this->buildElement('string', 1, ['max' => 10]),
+            'postal_code' => $this->buildElement('string', 1, ['max' => 16]),
             'state' => new Assert\Optional(
-                $this->buildElement('string', 0, ['max'=>14])
+                $this->buildElement('string', 0, ['max' => 14])
             ),
         ]);
     }
 
-    private function getCardHolderInfoConstraints()
+    private function getCardHolderInfoConstraints(): object
     {
         return new Assert\Collection([
             'email_address' => new Assert\Optional(
-                new Assert\Email(['mode'=>'loose'])
+                new Assert\Email(['mode' => 'loose'])
             ),
             'mobile_phone_number' => new Assert\Optional($this->buildElement('string')),
             'work_phone_number' => new Assert\Optional($this->buildElement('string')),
             'home_phone_number' => new Assert\Optional($this->buildElement('string')),
-        ]); 
+        ]);
     }
 
-    private function buildElement($typeValue, bool $isRequired=false, $length=0, $args=0)
+    private function buildElement(
+        string $typeValue,
+        bool $isRequired = false,
+        array $length = null,
+        $args = null // TODO can it be null?
+    ): object
     {
-        $inside_array = [
+        $inside_array = $this->getInsideArray($typeValue);
+        if ($isRequired) array_unshift($inside_array, new Assert\NotBlank());
+        if ($length) array_push($inside_array, new Assert\Length($length));
+        if ($args) array_push($inside_array, $args);
+
+        return $isRequired
+            ? new Assert\Required($inside_array)
+            : new Assert\Optional($inside_array)
+        ;
+    }
+
+    private function getInsideArray(string $typeValue): array
+    {
+        return [
             new Assert\Type([
                 'type' => $typeValue,
                 'message' => 'The value {{ value }} is not a valid {{ type }}.'
             ]),
         ];
-        if ($isRequired) array_unshift($inside_array, new Assert\NotBlank());
-        if ($length) array_push($inside_array, new Assert\Length($length));
-        if ($args) array_push($inside_array, $args);
-        
-        return $isRequired 
-            ? new Assert\Required($inside_array)
-            : new Assert\Optional($inside_array)
-        ;
     }
 }