+++ tests/ClientTestCase.php
@@ -10,7 +10,10 @@ use PHPUnit\Framework\TestCase;
 
 class ClientTestCase extends TestCase
 {
-    public function setUp() : void
+    /**
+     * @return void
+     */
+    public function setUp(): void
     {
         $log = Client::LOG_NONE;
 
@@ -19,6 +22,9 @@ class ClientTestCase extends TestCase
         $this->assertInstanceOf('Cardinity\Client', $this->client);
     }
 
+    /**
+     * @return array
+     */
     protected function getConfig()
     {
         return [
@@ -27,6 +33,9 @@ class ClientTestCase extends TestCase
         ];
     }
 
+    /**
+     * @return array
+     */
     protected function getPaymentParams()
     {
         return [
@@ -47,6 +56,23 @@ class ClientTestCase extends TestCase
         ];
     }
 
+    /**
+     * @param array optional
+     * @return array
+     */
+    public function get3ds2PaymentParams($browserData = [])
+    {
+        return array_merge(
+            $this->getPaymentParams(),
+            [
+                'threeds2_data' => $this->getThreeDS2Data($browserData)
+            ]
+        );
+    }
+
+    /**
+     * @return Payment\Payment
+     */
     public function getPayment()
     {
         $payment = new Payment\Payment();
@@ -58,7 +84,10 @@ class ClientTestCase extends TestCase
         return $payment;
     }
 
-    
+    /**
+     * @param array optional
+     * @return array
+     */
     public function getBrowserInfo($args = [])
     {
         $browserInfo = [
@@ -71,18 +100,18 @@ class ClientTestCase extends TestCase
             'color_depth' => 24,
             'time_zone' => -60,
         ];
-        if($args && isset($args['ip_address'])) {
-            $browserInfo['ip_address'] = $args['ip_address'];
-        }
-        if($args && isset($args['javascript_enabled'])) {
-            $browserInfo['javascript_enabled'] = $args['javascript_enabled'];
-        }
-        if($args && isset($args['java_enabled'])) {
-            $browserInfo['java_enabled'] = $args['java_enabled'];
+        if ($args) {
+            foreach($args as $key => $val) {
+                $browserInfo[$key] = $val;
+            }
         }
         return $browserInfo;
     }
 
+    /**
+     * @param array optional
+     * @return array
+     */
     public function getAddress($args = [])
     {
         $address = [
@@ -91,28 +120,33 @@ class ClientTestCase extends TestCase
             'country' => 'LT',
             'postal_code' => '0234'
         ];
-        if($args && isset($args['address_line2'])) {
+        if ($args && isset($args['address_line2'])) {
             $address['address_line2'] = $args['address_line2'];
         }
-        if($args && isset($args['address_line3'])) {
+        if ($args && isset($args['address_line3'])) {
             $address['address_line3'] = $args['address_line3'];
         }
-        if($args && isset($args['state'])) {
+        if ($args && isset($args['state'])) {
             $address['state'] = $args['state'];
         }
         return $address;
     }
 
-    public function getThreeDS2DataMandatoryData()
+    /**
+     * @param array optional
+     * @return array
+     */
+    public function getThreeDS2Data($browserData = [])
     {
-        $treeDS2Data['notification_url'] = 'https://notification.url/';
-
-        $browserInfo = $this->getBrowserInfo();
-        $treeDS2Data['browser_info'] = $browserInfo;
-
-        return $treeDS2Data;
+        return [
+            'notification_url' => 'https://notification.url/',
+            'browser_info' => $this->getBrowserInfo($browserData),
+        ];
     }
 
+    /**
+     * @return Payment\PaymentInstrumentCard
+     */
     public function getCard()
     {
         $card = new Payment\PaymentInstrumentCard();