+++ tests/SettlementTest.php
@@ -7,6 +7,9 @@ use Cardinity\Method\Payment;
 
 class SettlementTest extends ClientTestCase
 {
+    /**
+     * @return void
+     */
     public function testResultObjectSerialization()
     {
         $refund = new Settlement\Settlement();
@@ -21,6 +24,9 @@ class SettlementTest extends ClientTestCase
         );
     }
 
+    /**
+     * @return void
+     */
     public function testResultObjectUnserialization()
     {
         $json = '{"id":"foo","amount":"55.00","type":"bar"}';
@@ -35,7 +41,7 @@ class SettlementTest extends ClientTestCase
     }
 
     /**
-     * @return Payment\Payment
+     * @return Cardinity\Method\ResultObject $payment
      */
     public function testCreatePayment()
     {
@@ -47,7 +53,8 @@ class SettlementTest extends ClientTestCase
 
     /**
      * @depends testCreatePayment
-     * @expectedException Cardinity\Exception\Declined
+     * @param Payment\Payment $payment
+     * @return void
      */
     public function testCreateFail(Payment\Payment $payment)
     {
@@ -56,11 +63,14 @@ class SettlementTest extends ClientTestCase
             10.00,
             'fail'
         );
+        $this->expectException(\Cardinity\Exception\Declined::class);
         $this->client->call($method);
     }
 
     /**
      * @depends testCreatePayment
+     * @param Payment\Payment $payment
+     * @return Cardinity\Method\ResultObject $result
      */
     public function testCreate(Payment\Payment $payment)
     {
@@ -80,6 +90,8 @@ class SettlementTest extends ClientTestCase
 
     /**
      * @depends testCreate
+     * @param Settlement\Settlement $settlement
+     * @return void
      */
     public function testGet(Settlement\Settlement $settlement)
     {
@@ -98,6 +110,8 @@ class SettlementTest extends ClientTestCase
 
     /**
      * @depends testCreate
+     * @param Settlement\Settlement $settlement
+     * @return void
      */
     public function testGetAll(Settlement\Settlement $settlement)
     {