+++ tests/ThreeDS2Test.php
@@ -36,9 +36,9 @@ class ThreeDS2Test extends ClientTestCase
     }
 
     public function testClientCallSuccess()
-    {        
-        $threeDS2Data = $this->getThreeDS2DataMandatoryData();
-        unset($threeDS2Data['notification_url']);
+    {
+
+        $threeDS2Data = $this->getThreeDS2Data();
 
         $method = new Payment\Create([
             'amount' => 59.01,
@@ -62,67 +62,16 @@ class ThreeDS2Test extends ClientTestCase
             $payment = $this->client->call($method);
             $this->assertEquals('pending', $payment->getStatus());
         } catch (Exception\Declined $exception) {
-            /** @type Cardinity\Method\Payment\Payment */
             $payment = $exception->getResult();
-            $status = $payment->getStatus(); // value will be 'declined'
-            $errors = $exception->getErrors(); // list of errors occurred
+            $errors = $exception->getErrors();
         } catch (Exception\ValidationFailed $exception) {
-            /** @type Cardinity\Method\Payment\Payment */
             $payment = $exception->getResult();
-            $status = $payment->getStatus(); // value will be 'declined'
-            $errors = $exception->getErrors(); // list of errors occurred
-        } catch (Cardinity\Exception\InvalidAttributeValue $exception) {
-            /** @type Cardinity\Method\Payment\Payment */
-            // $payment = $exception->getResult();
-            // $status = $payment->getStatus(); // value will be 'declined'
-            $errors = $exception->getErrors(); // list of errors occurred
+            $errors = $exception->getErrors();
+        } catch (Exception\InvalidAttributeValue $exception) {
+            $errors = $exception->getViolations();
         }
-    }
-
-    public function ThreeDS2PaymentProvider()
-    {
-        $paymentArr = $this->getPaymentParams();
-        $paymentArr['threeds2_data'] = $this->ThreeDS2DataProvider();
-        return $paymentArr;
-    }
-    
-    public function ThreeDS2DataProvider(array $args = [])
-    {
-        $data = [
-            "notification_url" => "https://notification.url/",
-            "browser_info"=> $this->browserInfoProvider()
-        ];
-        if ($args) array_push($data, $args);
-
-        return $data;
-    }
-
-    public function addressProvider(array $args = [])
-    {
-        $address = [
-            "address_line1"=>"adress 1",
-            "city"=>"city",
-            "country"=>"LT",
-            "postal_code"=>"02245"
-        ];
-        if ($args) foreach ($args as $k => $v) $address[$k] = $v;
-
-        return $address;
-    }
-
-    public function browserInfoProvider(array $args = [])
-    {
-        $info = [
-            "accept_header"=>"HTTP accept header.",
-            "browser_language"=>"cardholder language IETF BCP 47.",
-            "screen_width"=>600,
-            "screen_height"=>400,
-            "challenge_window_size"=>"600x400",
-            "user_agent"=>"agent James Bond",
-            "color_depth"=>24,
-            "time_zone"=>-60
-        ];
-        if ($args) foreach ($args as $k => $v) $info[$k] = $v;
-        return $info;
+        if (isset($errors)) {
+            $this->assertContains('[threeds2_data][notification_url]',$errors);
+        }
     }
 }